return npmerr globally
